[Refactor] 리뷰 비드백 반영 및 소비자 클레임 도메인 통합
