+++ src/main/java/com/bbangle/bbangle/claim/customer/controller/CustomerOrderController.java
@@ -1,7 +1,9 @@
 package com.bbangle.bbangle.claim.customer.controller;
 
+import com.bbangle.bbangle.claim.customer.controller.dto.CustomerCancelRequest;
 import com.bbangle.bbangle.claim.customer.controller.dto.CustomerExchangeRequest;
-import com.bbangle.bbangle.claim.customer.service.CustomerExchangeService;
+import com.bbangle.bbangle.claim.customer.controller.dto.CustomerReturnRequest;
+import com.bbangle.bbangle.claim.customer.service.CustomerOrderService;
 import com.bbangle.bbangle.common.dto.CommonResult;
 import com.bbangle.bbangle.common.service.ResponseService;
 import com.bbangle.bbangle.config.security.CustomerApiPath;
@@ -17,10 +19,30 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @RequestMapping(CustomerApiPath.PREFIX + "/orders")
 @RestController
-public class CustomerExchangeController {
+public class CustomerOrderController {
 
     private final ResponseService responseService;
-    private final CustomerExchangeService customerExchangeService;
+    private final CustomerOrderService customerOrderService;
+
+    @PatchMapping("/{orderId}/cancel")
+    public CommonResult requestCancel(
+        @PathVariable Long orderId,
+        @Valid @RequestBody CustomerCancelRequest request,
+        @AuthenticationPrincipal Long customerId
+    ) {
+        customerOrderService.cancelOrder(orderId, customerId, request);
+        return responseService.getSuccessResult();
+    }
+
+    @PatchMapping("/{orderId}/return")
+    public CommonResult requestReturn(
+        @PathVariable Long orderId,
+        @Valid @RequestBody CustomerReturnRequest request,
+        @AuthenticationPrincipal Long customerId
+    ) {
+        customerOrderService.requestReturn(orderId, customerId, request);
+        return responseService.getSuccessResult();
+    }
 
     @PatchMapping("/{orderId}/exchange")
     public CommonResult requestExchange(
@@ -28,7 +50,7 @@ public class CustomerExchangeController {
         @Valid @RequestBody CustomerExchangeRequest request,
         @AuthenticationPrincipal Long customerId
     ) {
-        customerExchangeService.requestExchange(orderId, customerId, request);
+        customerOrderService.requestExchange(orderId, customerId, request);
         return responseService.getSuccessResult();
     }
 }
